deploy: 903b5324f8c90157acb86338acd872044e7e56be

+++ js/fermeture.js
@@ -10,6 +10,9 @@ function checkClosed() {
 checkClosed();
 
 function displayClosedOverlay() {
+  if (window.location.search === "?DEBUG") {
+    return;
+  }
   window.localStorage && window.localStorage.clear();
 
   document.body.innerHTML = `
